+++ GameServer/packets/Client/168/UseSlotHandler.cs
@@ -84,11 +84,11 @@ namespace DOL.GS.PacketHandler.Client.v168
 				var player = (GamePlayer) m_actionSource;
 				if ((m_flagSpeedData & 0x200) != 0)
 				{
-					player.CurrentSpeed = -(m_flagSpeedData & 0x1ff); // backward movement
+					player.CurrentSpeed = (short)(-(m_flagSpeedData & 0x1ff)); // backward movement
 				}
 				else
 				{
-					player.CurrentSpeed = m_flagSpeedData & 0x1ff; // forwardmovement
+					player.CurrentSpeed = (short)(m_flagSpeedData & 0x1ff); // forwardmovement
 				}
 				player.IsStrafing = (m_flagSpeedData & 0x4000) != 0;
 				player.TargetInView = (m_flagSpeedData & 0xa000) != 0; // why 2 bits? that has to be figured out
