+++ GameServer/spells/Heretic/HereticSpeedDecrease.cs
@@ -210,7 +210,7 @@ namespace DOL.GS.Spells
 			GameNPC npc = owner as GameNPC;
 			if (npc != null)
 			{
-				int maxSpeed = npc.MaxSpeed;
+				short maxSpeed = npc.MaxSpeed;
 				if (npc.CurrentSpeed > maxSpeed)
 					npc.CurrentSpeed = maxSpeed;
 			}
